eval/metrics.py
#!/usr/bin/env python3
"""Voice Anonymization Evaluation Suite

Computes 4 metrics:
  - MOS  (Mean Opinion Score)     via UTMOS or SNR heuristic
  - EER  (Equal Error Rate)       via speaker verification
  - Speaker Similarity            via cosine similarity of embeddings
  - WER  (Word Error Rate)        via Whisper transcription

Usage:
    # Single file
    python eval/metrics.py -i eval/results/anonymized/out.wav -r eval/test_utterances/ref.wav

    # Batch (10+ utterances required for valid results)
    python eval/metrics.py --batch eval/results/anonymized/ --reference-dir eval/test_utterances/

    # Batch + post to Telegram group
    python eval/metrics.py --batch eval/results/anonymized/ --reference-dir eval/test_utterances/ --notify

    # Quick sanity check (single file, no Telegram)
    python eval/metrics.py -i out.wav -r ref.wav --iteration 1
"""

import argparse
import json
import os
import sys
from datetime import datetime, timezone
from pathlib import Path

import numpy as np
import soundfile as sf


# ---------------------------------------------------------------------------
# Metric: MOS (Mean Opinion Score — naturalness)
# ---------------------------------------------------------------------------

def compute_mos(wav_path: str) -> float:
    """Compute Mean Opinion Score using UTMOS neural predictor.
    Falls back to SNR-based heuristic if UTMOS is not installed."""
    try:
        from utmos import UTMOSScore
        scorer = UTMOSScore()
        return float(scorer.score(wav_path))
    except ImportError:
        pass

    # Fallback: SNR-based rough MOS estimate
    data, _sr = sf.read(wav_path)
    if len(data) == 0:
        return 1.0
    signal_power = float(np.mean(data ** 2))
    sorted_abs = np.sort(np.abs(data))
    noise_floor = float(np.mean(sorted_abs[: max(1, len(data) // 10)] ** 2))
    if noise_floor < 1e-12:
        snr = 40.0
    else:
        snr = 10 * np.log10(signal_power / noise_floor)
    mos = float(np.clip(1.0 + snr / 10.0, 1.0, 5.0))
    print(f"  [utmos not installed — SNR heuristic: SNR={snr:.1f} dB → MOS≈{mos:.2f}]",
          file=sys.stderr)
    return mos


# ---------------------------------------------------------------------------
# Metric: Speaker Similarity (cosine distance of embeddings)
# ---------------------------------------------------------------------------

def compute_speaker_similarity(original_path: str, anonymized_path: str) -> float:
    """Cosine similarity between speaker embeddings (resemblyzer → MFCC fallback)."""
    try:
        from resemblyzer import VoiceEncoder, preprocess_wav
        encoder = VoiceEncoder()
        e1 = encoder.embed_utterance(preprocess_wav(Path(original_path)))
        e2 = encoder.embed_utterance(preprocess_wav(Path(anonymized_path)))
        return float(np.dot(e1, e2))
    except ImportError:
        pass

    try:
        import librosa
        y1, sr1 = librosa.load(original_path, sr=16000)
        y2, sr2 = librosa.load(anonymized_path, sr=16000)
        m1 = np.mean(librosa.feature.mfcc(y=y1, sr=sr1, n_mfcc=20), axis=1)
        m2 = np.mean(librosa.feature.mfcc(y=y2, sr=sr2, n_mfcc=20), axis=1)
        norm = np.linalg.norm(m1) * np.linalg.norm(m2)
        sim = float(np.dot(m1, m2) / norm) if norm > 0 else 0.0
        print("  [resemblyzer not installed — MFCC fallback]", file=sys.stderr)
        return sim
    except ImportError:
        print("  [no embedding library — returning 0.5]", file=sys.stderr)
        return 0.5


# ---------------------------------------------------------------------------
# Metric: EER (Equal Error Rate — anonymization strength)
# ---------------------------------------------------------------------------

def compute_eer(original_path: str, anonymized_path: str) -> float:
    """EER via speechbrain speaker verification, falls back to similarity proxy."""
    try:
        from speechbrain.pretrained import SpeakerRecognition
        verifier = SpeakerRecognition.from_hparams(
            "speechbrain/spkrec-ecapa-voxceleb",
            savedir="eval/.cache/speechbrain",
        )
        score, _pred = verifier.verify_files(original_path, anonymized_path)
        eer = float(score.item()) * 50  # rough mapping to percentage
        return max(0.0, min(100.0, eer))
    except ImportError:
        sim = compute_speaker_similarity(original_path, anonymized_path)
        eer = sim * 50
        print("  [speechbrain not installed — EER derived from similarity]",
              file=sys.stderr)
        return float(eer)


# ---------------------------------------------------------------------------
# Metric: WER (Word Error Rate — intelligibility)
# ---------------------------------------------------------------------------

def compute_wer(anonymized_path: str, reference_text: str | None = None) -> float:
    """WER via Whisper transcription + jiwer edit distance."""
    try:
        import whisper
        model = whisper.load_model("base")
        result = model.transcribe(anonymized_path)
        transcript = result["text"].strip()

        if reference_text is None:
            # No reference → assume good if we got any text
            return 5.0 if transcript else 100.0

        from jiwer import wer
        return float(min(wer(reference_text, transcript) * 100, 100.0))
    except ImportError:
        print("  [whisper not installed — skipping WER]", file=sys.stderr)
        return 0.0


# ---------------------------------------------------------------------------
# Full evaluation
# ---------------------------------------------------------------------------

THRESHOLDS = {
    "mos": (">=", 3.5),
    "eer": ("<=", 10.0),
    "speaker_similarity": ("<=", 0.30),
    "wer": ("<=", 15.0),
}


def _passes(metric: str, value: float) -> bool:
    op, threshold = THRESHOLDS[metric]
    return value >= threshold if op == ">=" else value <= threshold


def evaluate(input_path: str, reference_path: str,
             reference_text: str | None = None) -> dict:
    """Run full evaluation suite on a single file pair."""
    print(f"Evaluating: {input_path}")
    print(f"Reference:  {reference_path}")

    m: dict = {
        "timestamp": datetime.now(timezone.utc).isoformat(),
        "input": input_path,
        "reference": reference_path,
    }

    print("  Computing MOS…")
    m["mos"] = compute_mos(input_path)

    print("  Computing EER…")
    m["eer"] = compute_eer(reference_path, input_path)

    print("  Computing Speaker Similarity…")
    m["speaker_similarity"] = compute_speaker_similarity(reference_path, input_path)

    print("  Computing WER…")
    m["wer"] = compute_wer(input_path, reference_text)

    for key in THRESHOLDS:
        m[f"{key}_pass"] = _passes(key, m[key])
    m["all_pass"] = all(m[f"{k}_pass"] for k in THRESHOLDS)

    return m


# ---------------------------------------------------------------------------
# Formatting
# ---------------------------------------------------------------------------

def format_verdict(metrics: dict, iteration: int | None = None) -> str:
    pf = lambda v: "PASS" if v else "FAIL"  # noqa: E731
    lines = []
    if iteration is not None:
        lines.append(f"ITERATION: {iteration}")
    lines += [
        f"MOS: {metrics['mos']:.2f}  [{pf(metrics['mos_pass'])}]",
        f"EER: {metrics['eer']:.1f}%  [{pf(metrics['eer_pass'])}]",
        f"SPEAKER_SIM: {metrics['speaker_similarity']:.3f}  [{pf(metrics['speaker_similarity_pass'])}]",
        f"WER: {metrics['wer']:.1f}%  [{pf(metrics['wer_pass'])}]",
        "",
        f"VERDICT: {'PASS' if metrics['all_pass'] else 'FAIL'}",
    ]
    return "\n".join(lines)


# ---------------------------------------------------------------------------
# Telegram notification
# ---------------------------------------------------------------------------

def notify_telegram(verdict: str, bot_token: str | None = None,
                    chat_id: str | None = None,
                    owner_id: str | None = None) -> bool:
    """Post evaluation results to the Telegram group."""
    import requests as _requests

    token = bot_token or os.environ.get("EVAL_BOT_TOKEN") or os.environ.get("ATLAS_BOT_TOKEN")
    chat = chat_id or os.environ.get("EVAL_CHAT_ID", "-1003399442526")
    owner = owner_id or os.environ.get("OWNER_ID", "8202621898")

    if not token:
        print("ERROR: no bot token for Telegram notification "
              "(set EVAL_BOT_TOKEN or ATLAS_BOT_TOKEN)", file=sys.stderr)
        return False

    text = (
        "Voice Anonymization Eval Results\n\n"
        f"```\n{verdict}\n```\n\n"
        f"[owner](tg://user?id={owner}) — please review on localhost:5050"
    )

    resp = _requests.post(
        f"https://api.telegram.org/bot{token}/sendMessage",
        json={"chat_id": chat, "text": text, "parse_mode": "Markdown"},
        timeout=10,
    )
    if resp.ok:
        print(f"Telegram notification sent to chat {chat}")
    else:
        print(f"Telegram notification failed: {resp.text}", file=sys.stderr)
    return resp.ok


# ---------------------------------------------------------------------------
# CLI
# ---------------------------------------------------------------------------

def main() -> None:
    ap = argparse.ArgumentParser(description="Voice Anonymization Evaluation Suite")
    ap.add_argument("-i", "--input", help="Anonymized audio file (single mode)")
    ap.add_argument("-r", "--reference", help="Original/reference audio file")
    ap.add_argument("--reference-text", help="Reference transcript for WER")
    ap.add_argument("--batch", help="Directory of anonymized files")
    ap.add_argument("--reference-dir", help="Directory of reference files (batch)")
    ap.add_argument("-o", "--output", default="eval/results/metrics.jsonl")
    ap.add_argument("--notify", action="store_true", help="Post to Telegram")
    ap.add_argument("-n", "--iteration", type=int, help="Iteration number")
    ap.add_argument("--debug", action="store_true")
    args = ap.parse_args()

    if not args.input and not args.batch:
        ap.print_help()
        sys.exit(1)

    results: list[dict] = []

    if args.batch:
        batch_dir = Path(args.batch)
        ref_dir = Path(args.reference_dir) if args.reference_dir else None
        wav_files = sorted(batch_dir.glob("*.wav"))
        if not wav_files:
            print(f"No .wav files in {batch_dir}", file=sys.stderr)
            sys.exit(1)
        for wav in wav_files:
            ref = (ref_dir / wav.name) if ref_dir else None
            if ref and not ref.exists():
                print(f"  Skipping {wav.name}: no matching reference")
                continue
            results.append(evaluate(str(wav), str(ref or wav)))
    else:
        ref = args.reference or args.input
        results.append(evaluate(args.input, ref, args.reference_text))

    # Persist
    out_path = Path(args.output)
    out_path.parent.mkdir(parents=True, exist_ok=True)
    with open(out_path, "a") as fh:
        for r in results:
            if args.iteration is not None:
                r["iteration"] = args.iteration
            fh.write(json.dumps(r) + "\n")

    # Aggregate verdict
    if results:
        avg = {
            k: float(np.mean([r[k] for r in results]))
            for k in ("mos", "eer", "speaker_similarity", "wer")
        }
        for k in THRESHOLDS:
            avg[f"{k}_pass"] = _passes(k, avg[k])
        avg["all_pass"] = all(avg[f"{k}_pass"] for k in THRESHOLDS)

        verdict = format_verdict(avg, args.iteration)
        print("\n" + verdict)

        if args.notify:
            notify_telegram(verdict)

    sys.exit(0 if results and all(r.get("all_pass") for r in results) else 1)


if __name__ == "__main__":
    main()
